Create the EventFactory
